+++ utils/meta/index.js
@@ -4,11 +4,16 @@
 
 const preMarkup = require('@minna-ui/pre-markup');
 const preStyle = require('@minna-ui/pre-style');
-const { catchErr, makeCss, makeHtml } = require('@minna-ui/rollup-plugins');
+const {
+  catchErr,
+  gitDescribe,
+  makeCss,
+  makeHtml,
+} = require('@minna-ui/rollup-plugins');
 
 const prod = process.env.NODE_ENV === 'production';
 
-/** Svelte preprocessors preset. */
+/** Svelte preprocessor preset. */
 const preprocess = {
   markup: preMarkup({ level: prod ? 2 : 0 }),
   style: preStyle(),
@@ -16,6 +21,7 @@ const preprocess = {
 
 module.exports = {
   catchErr,
+  gitDescribe,
   makeCss,
   makeHtml,
   preMarkup,