+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- New `gitDescribe` helper function in `@minna-ui/rollup-plugins`.
+
 ## [0.18.1] - 2018-12-19
 
 ### Added